perf(fonts): optimize font styles
Refactored AppText component styles, so that they can become more
suitable for most of the fonts in the application.

Was removed styles.js file from config folder due to the fact that there
is no need for it.

Was optimized styles and structure for a bunch of components to make
them more clearness and readable:
* AppButton - Text component replaced by AppText. In styles.text was
removed useless props related to font styles.

* ErrorMessage - changed like AppButton.

* AppTextInput - added fontFamily property to styles.input.

Closes: #20