feat(AppTextInput): redesign for android
AppTextInput got two specialized implementations for ios and android.
The old one migrated to the implementation for the iPhone and the other
was designed from scratch.

formFieldsProps.js config file was also split into two files: one for
iso and others for android, so that thay can contain more flexible and
suitable properties for AppTextInput arhitecture and plaform.

Added new color to palette into colors.js config file. It intended for
android AppTextInput border.

Modified PasswordCreateScreen AppTextInput props, so that depending on
the platform they provide different kinds of properties.

Closes: #61